+++ internal/daemon/controller/handlers/credentials/credential_service.go
@@ -360,6 +360,9 @@ func (s Service) updateInRepo(
 	switch subtypes.SubtypeFromId(domain, id) {
 	case static.UsernamePasswordSubtype:
 		cred, err := toUsernamePasswordStorageCredential(storeId, in)
+		if err != nil {
+			return nil, errors.Wrap(ctx, err, op, errors.WithMsg("unable to convert to username/password storage credential"))
+		}
 		cred.PublicId = id
 		repo, err := s.repoFn()
 		if err != nil {
@@ -530,7 +533,7 @@ func toUsernamePasswordStorageCredential(storeId string, in *pb.Credential) (out
 //  * All required parameters are set
 //  * There are no conflicting parameters provided
 func validateGetRequest(req *pbs.GetCredentialRequest) error {
-	return handlers.ValidateGetRequest(handlers.NoopValidatorFn, req, static.CredentialPrefix)
+	return handlers.ValidateGetRequest(handlers.NoopValidatorFn, req, static.UsernamePasswordCredentialPrefix, static.PreviousUsernamePasswordCredentialPrefix)
 }
 
 func validateCreateRequest(req *pbs.CreateCredentialRequest) error {
@@ -539,7 +542,7 @@ func validateCreateRequest(req *pbs.CreateCredentialRequest) error {
 		if req.Item.GetType() != static.UsernamePasswordSubtype.String() {
 			badFields[globals.TypeField] = fmt.Sprintf("Unsupported credential type %q", req.Item.GetType())
 		}
-		if !handlers.ValidId(handlers.Id(req.Item.GetCredentialStoreId()), static.CredentialStorePrefix) {
+		if !handlers.ValidId(handlers.Id(req.Item.GetCredentialStoreId()), static.CredentialStorePrefix, static.PreviousCredentialStorePrefix) {
 			badFields[globals.CredentialStoreIdField] = "This field must be a valid credential store id."
 		}
 
@@ -570,16 +573,16 @@ func validateUpdateRequest(req *pbs.UpdateCredentialRequest) error {
 		}
 
 		return badFields
-	}, static.CredentialPrefix)
+	}, static.UsernamePasswordCredentialPrefix, static.PreviousUsernamePasswordCredentialPrefix)
 }
 
 func validateDeleteRequest(req *pbs.DeleteCredentialRequest) error {
-	return handlers.ValidateDeleteRequest(handlers.NoopValidatorFn, req, static.CredentialPrefix)
+	return handlers.ValidateDeleteRequest(handlers.NoopValidatorFn, req, static.UsernamePasswordCredentialPrefix, static.PreviousUsernamePasswordCredentialPrefix)
 }
 
 func validateListRequest(req *pbs.ListCredentialsRequest) error {
 	badFields := map[string]string{}
-	if !handlers.ValidId(handlers.Id(req.GetCredentialStoreId()), static.CredentialStorePrefix) {
+	if !handlers.ValidId(handlers.Id(req.GetCredentialStoreId()), static.CredentialStorePrefix, static.PreviousCredentialStorePrefix) {
 		badFields[globals.CredentialStoreIdField] = "This field must be a valid credential store id."
 	}
 	if _, err := handlers.NewFilter(req.GetFilter()); err != nil {