+++ internal/daemon/controller/handlers/credentials/credential_service_test.go
@@ -172,6 +172,7 @@ func TestGet(t *testing.T) {
 
 	store := static.TestCredentialStore(t, conn, wrapper, prj.GetPublicId())
 	cred := static.TestUsernamePasswordCredential(t, conn, wrapper, "user", "pass", store.GetPublicId(), prj.GetPublicId())
+	credPrev := static.TestUsernamePasswordCredential(t, conn, wrapper, "user", "pass", store.GetPublicId(), prj.GetPublicId(), static.WithPublicId(fmt.Sprintf("%s_1234567890", static.PreviousUsernamePasswordCredentialPrefix)))
 	s, err := NewService(staticRepoFn, iamRepoFn)
 	require.NoError(t, err)
 
@@ -210,8 +211,30 @@ func TestGet(t *testing.T) {
 			},
 		},
 		{
+			name: "success-prev-prefix",
+			id:   credPrev.GetPublicId(),
+			res: &pbs.GetCredentialResponse{
+				Item: &pb.Credential{
+					Id:                credPrev.GetPublicId(),
+					CredentialStoreId: cred.GetStoreId(),
+					Scope:             &scopepb.ScopeInfo{Id: store.GetScopeId(), Type: scope.Project.String(), ParentScopeId: prj.GetParentId()},
+					Type:              static.UsernamePasswordSubtype.String(),
+					AuthorizedActions: testAuthorizedActions,
+					CreatedTime:       credPrev.CreateTime.GetTimestamp(),
+					UpdatedTime:       credPrev.UpdateTime.GetTimestamp(),
+					Version:           1,
+					Attrs: &pb.Credential_UsernamePasswordAttributes{
+						UsernamePasswordAttributes: &pb.UsernamePasswordAttributes{
+							Username:     wrapperspb.String("user"),
+							PasswordHmac: base64.RawURLEncoding.EncodeToString([]byte(hm)),
+						},
+					},
+				},
+			},
+		},
+		{
 			name: "not found error",
-			id:   fmt.Sprintf("%s_1234567890", static.CredentialPrefix),
+			id:   fmt.Sprintf("%s_1234567890", static.UsernamePasswordCredentialPrefix),
 			err:  handlers.NotFoundError(),
 		},
 		{
@@ -279,7 +302,7 @@ func TestDelete(t *testing.T) {
 		},
 		{
 			name: "not found error",
-			id:   fmt.Sprintf("%s_1234567890", static.CredentialPrefix),
+			id:   fmt.Sprintf("%s_1234567890", static.UsernamePasswordCredentialPrefix),
 			err:  handlers.NotFoundError(),
 		},
 		{
@@ -334,7 +357,7 @@ func TestCreate(t *testing.T) {
 			name: "Can't specify Id",
 			req: &pbs.CreateCredentialRequest{Item: &pb.Credential{
 				CredentialStoreId: store.GetPublicId(),
-				Id:                static.CredentialPrefix + "_notallowed",
+				Id:                static.UsernamePasswordCredentialPrefix + "_notallowed",
 				Type:              static.UsernamePasswordSubtype.String(),
 				Attrs: &pb.Credential_UsernamePasswordAttributes{
 					UsernamePasswordAttributes: &pb.UsernamePasswordAttributes{
@@ -446,9 +469,9 @@ func TestCreate(t *testing.T) {
 					},
 				},
 			}},
-			idPrefix: static.CredentialPrefix + "_",
+			idPrefix: static.UsernamePasswordCredentialPrefix + "_",
 			res: &pbs.CreateCredentialResponse{
-				Uri: fmt.Sprintf("credentials/%s_", static.CredentialPrefix),
+				Uri: fmt.Sprintf("credentials/%s_", static.UsernamePasswordCredentialPrefix),
 				Item: &pb.Credential{
 					Id:                store.GetPublicId(),
 					CredentialStoreId: store.GetPublicId(),